pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prototype-hls"
version = "0.1.0"
description = "Synthetic Outpatient Flow Analytics Demo for OpenShift"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
]

[project.optional-dependencies]
db = ["psycopg2-binary>=2.9"]
ml = [
    "scikit-learn>=1.3",
    "xgboost>=2.0",
]
gpu = [
    "cudf-cu12",
    "cuml-cu12",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
all = [
    "prototype-hls[db,ml,dev]",
]

[project.scripts]
hls-generate = "src.generator.generate:main"
hls-etl = "src.etl.batch_etl:main"
hls-analytics = "src.analytics.analytics:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
